Framework for Items
Changed update/draw system to be two generic lists.  Item framework is
in place, skeletal for JSON later.

+++ core/src/com/redtoorange/delver/utility/Dice.java
@@ -13,7 +13,7 @@ public class Dice {
     private int sides;
     private int count;
 
-    public float maxDamage(){
+    public int maxDamage(){
         return sides * count;
     }
